JUnit: Update Test
